+++ msb/opt/video/WebCamSource.py
